Hierarchical Density-Based Spatial Clustering of Applications with Noise

Hierarchical Density-Based Spatial Clustering of Applications with Noise

# S3 method for Seurat
HDBSCAN(
  object,
  reduction = "umap",
  dims = NULL,
  algorithm = "best",
  alpha = 1,
  prediction_data = TRUE,
  approx_min_span_tree = TRUE,
  gen_min_span_tree = FALSE,
  leaf_size = 30,
  metric = "euclidean",
  min_cluster_size = 1000L,
  min_samples = 1L,
  cluster_selection_epsilon = 0.1,
  cluster_selection_method = "leaf",
  nThreads = parallel::detectCores() - 1,
  return_seurat = TRUE,
  return_full = FALSE
)

Arguments

object

Seurat object

reduction

name of embedding to use (default='umap')

dims

number of dimensions to use from the reduction embedding. If not set (default=NULL), uses all available dimensions.

algorithm

specify which algorithm to use

alpha

A distance scaling parameter as used in robust single linkage.

prediction_data

not sure what this is for. Will update later.

approx_min_span_tree

Whether to accept an only approximate minimum spanning tree.

gen_min_span_tree

Whether to generate the minimum spanning tree with regard to mutual reachability distance for later analysis.

leaf_size

If using a space tree algorithm (kdtree, or balltree) the number of points ina leaf node of the tree.

metric

The metric to use when calculating distance between instances in a feature array.

min_cluster_size

The minimum size of clusters

min_samples

The number of samples in a neighbourhood for a point to be considered a core point.

cluster_selection_epsilon

A distance threshold. Clusters below this value will be merged.

cluster_selection_method

The method used to select clusters from the condensed tree.

nThreads

number of parallel threads

return_seurat

logical to return the result within the orignal object or as the raw HDBSCAN result

return_full

return the original full python result

Value

returns the result from HDBSCAN.matrix or the original Seurat object with the result from HDBSCAN.matrix stored in the misc slot.

Examples

if (FALSE) {
# load Seurat object
data('pbmc_small', package = 'SeuratObject')

# tsne embedding matrix
emb <- Seurat::Embeddings(object = pbmc_small, 'tsne')

# HDBSCAN.matrix
res.matrix <- dbsinglecell::HDBSCAN(emb, min_cluster_size = 5)
res.matrix

# HDBSCAN.Seurat
res.seurat <- dbsinglecell::HDBSCAN(pbmc_small,reduction = 'tsne', min_cluster_size = 5)
slot(res.seurat, 'misc')$hdbscan

}